implemented assassin invisibility skill (player only renders for assassin during the duration)
utililzed the boolean value to not draw for non assassin players. The skill only flips the enabled; after the duration ends on the client side, it automatically sends another invisibility skill packet to server to reverse it. Thus, you must always make sure that the cooldown is longer than the duration.